added benchmarkfile argument and debug flag, grayscott using 64 bit float instead of 32 bit l2norm uses a*a instead of a**2

+++ benchmarks/grayscott.py
@@ -6,8 +6,8 @@
 import numpy as np
 def grayscott(counts, Du, Dv, F, k):
     n = 300
-    U = np.zeros((n+2,n+2), dtype=np.float32)
-    V = np.zeros((n+2,n+2), dtype=np.float32)
+    U = np.zeros((n+2,n+2), dtype=np.float64)
+    V = np.zeros((n+2,n+2), dtype=np.float64)
     u, v = U[1:-1,1:-1], V[1:-1,1:-1]
 
     r = 20
